pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "cutie"
version = "1.0.0"
authors = [{ name = "Rex Cheng", email = "hkchengrex@gmail.com" }]
description = ""
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
dependencies = [
  'cython',
  'gitpython >= 3.1',
  'thinplate@git+https://github.com/cheind/py-thin-plate-spline',
  'hickle >= 5.0',
  'tensorboard >= 2.11',
  'numpy >= 1.21',
  'Pillow >= 9.5',
  'opencv-python >= 4.8',
  'scipy >= 1.7',
  'pycocotools >= 2.0.7',
  'tqdm >= 4.66.1',
  'gradio >= 3.34',
  'gdown >= 4.7.1',
  'einops >= 0.6',
  'hydra-core >= 1.3.2',
  'PySide6 >= 6.5.3',
  'charset-normalizer >= 3.1.0',
  'netifaces >= 0.11.0',
  'cchardet >= 2.1.7',
  'easydict',
  'av >= 0.5.2',
  'requests',
  'pyqtdarktheme',
]


[project.urls]
"Homepage" = "https://github.com/hkchengrex/Cutie"
"Bug Tracker" = "https://github.com/hkchengrex/Cutie/issues"
